+++ src/components/Zoom/Zoom.js
@@ -1,14 +1,38 @@
-import React, { useContext } from 'react';
+import React, { useContext, useEffect, useState } from 'react';
 import { Button, Slider, Flex } from 'theme-ui';
 import { FaPlus, FaMinus } from 'react-icons/fa';
+import numeral from 'numeral';
 import Context from '../../DefaultContext';
 import BaseComponent from '../_common/BaseComponent';
 import zoomIn from './util/zoomIn';
 import zoomOut from './util/zoomOut';
+import mapExists from '../../util/mapExists';
 
-const Zoom = ({ circular, horizontal, controlType, sx, ...rest }) => {
+const Zoom = ({
+  circular,
+  horizontal,
+  controlType,
+  showZoomLevel,
+  sx,
+  ...rest
+}) => {
   const config = useContext(Context);
   const { map } = config;
+  const [zoom, setZoom] = useState('');
+
+  function displayZoom() {
+    setZoom(numeral(map.getZoom()).format('0.0'));
+  }
+
+  useEffect(() => {
+    if (mapExists(map)) {
+      // map.on('load', () => {
+      map.on('zoom', () => {
+        displayZoom();
+      });
+      // });
+    }
+  }, [map]);
 
   let buttonStyle = {
     borderRadius: 'default',
@@ -32,6 +56,16 @@ const Zoom = ({ circular, horizontal, controlType, sx, ...rest }) => {
 
   if (horizontal) {
     flexDirection = 'row';
+  }
+
+  let zoomLevelIndicator = null;
+
+  if (showZoomLevel) {
+    zoomLevelIndicator = (
+      <Button sx={{ ...buttonStyle, bg: 'secondary', fontSize: 0 }}>
+        {zoom}
+      </Button>
+    );
   }
 
   const zoomInBtn = (
@@ -69,6 +103,7 @@ const Zoom = ({ circular, horizontal, controlType, sx, ...rest }) => {
       zoomControls = (
         <>
           {zoomOutBtn}
+          {zoomLevelIndicator}
           {zoomInBtn}
         </>
       );
@@ -76,6 +111,7 @@ const Zoom = ({ circular, horizontal, controlType, sx, ...rest }) => {
       zoomControls = (
         <>
           {zoomInBtn}
+          {zoomLevelIndicator}
           {zoomOutBtn}
         </>
       );